Se agrega validación para obtener información de ususario y demás datos

+++ domain/model/src/main/java/co/com/crediya/model/loantype/LoanType.java
@@ -5,6 +5,8 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @NoArgsConstructor
@@ -16,7 +18,7 @@ public class LoanType {
     String code;
     Double minimumAmount;
     Double maximumAmount;
-    Double interestRate;
+    BigDecimal interestRate;
     Boolean automaticValidation;
 
 }
